pyproject.toml
[project]
name = "weeaboo-buddy"
version = "0.1.0"
description = " An Agent to help those interested in Anime learn facts, trends, and statistics about the latest anime!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jikanpy-v4>=1.0.2",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.64",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "pymongo>=4.12.1",
    "rich>=14.0.0",
    "streamlit>=1.46.0",
]
